# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Buildspace API",
    "version": "0.1.0",
    "description": "Separate standalone backend contract. Full request schemas live alongside Hono routes in src/modules."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/api/v1/spaces/{slug}/calls": {
      "get": {
        "summary": "Read calls",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Write calls",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/spaces/{slug}/ai/settings": {
      "get": {
        "summary": "Read ai/settings",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "put": {
        "summary": "Write ai/settings",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/spaces/{slug}/ai/runs": {
      "get": {
        "summary": "Read ai/runs",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Write ai/runs",
        "tags": [
          "Realtime and AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
    "/api/v1/calls/{callId}/join": {
      "post": {
        "summary": "Issue LiveKit join token",
        "tags": [
          "Calls"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/calls/{callId}/end": {
      "post": {
        "summary": "End call",
        "tags": [
          "Calls"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/ai/runs/{runId}": {
      "get": {
        "summary": "AI run and citations",
        "tags": [
          "AI"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "buildspace_session"
      }
    }
  }
}